data_pull_tools/requirements.txt
httpx>=0.20.0
pydantic>=2.0.0
pydantic-ai>=0.1.0
python-dotenv>=1.0.0
selenium>=4.0.0
webdriver-manager>=3.8.0
youtube-search-python>=1.0.0
yt-dlp>=2023.01.01
huggingface-hub>=0.1.0
openai>=1.0.0